Add GitHub and Slack integration modules
Introduces src/github.ts and src/slack.ts for GitHub and Slack integration. Updates server.ts to utilize these modules. Migrates dependency management from npm (package-lock.json) to pnpm (pnpm-lock.yaml).
